Fix API to return numbers instead of strings for budget/ROI fields

+++ backend/campaigns/serializers.py
@@ -284,6 +284,6 @@ class CampaignStatsSerializer(serializers.Serializer):
     """Serializer for campaign statistics"""
     total_campaigns = serializers.IntegerField()
     active_campaigns = serializers.IntegerField()
-    total_budget = serializers.DecimalField(max_digits=10, decimal_places=2)
-    total_spent = serializers.DecimalField(max_digits=10, decimal_places=2)
-    average_roi = serializers.DecimalField(max_digits=5, decimal_places=2) 
+    total_budget = serializers.FloatField()
+    total_spent = serializers.FloatField()
+    average_roi = serializers.FloatField()
